Critical bug patch + more settings
QuickStart was setting rally point with a exec command which was pushing "unknow" resource gatherer if a unit was spawning on it after it would disappear. This bug wasn't occurring in normal circumstances because the resource is consumed slower then the time a player would change the rally point for the first time.